Refactor some ember data related core to a helper.
