+++ lib/formulas/ember-data-model-lookup-transform.js
@@ -6,6 +6,7 @@ var types      = recast.types.namedTypes;
 var Inflected  = require('inflected');
 var underscore = Inflected.underscore.bind(Inflected);
 var dasherize  = Inflected.dasherize.bind(Inflected);
+var emberData  = require('./helpers/ember-data');
 
 function normalize(string) {
   return dasherize(underscore(string));
@@ -14,93 +15,22 @@ function normalize(string) {
 module.exports = function transformEmberDataModelLookups(source){
   var ast = recast.parse(source);
 
-  var STORE_METHODS = [
-    'all',
-    'createRecord',
-    'filter',
-    'find',
-    'fetch',
-    'fetchAll',
-    'findById',
-    'findByIds',
-    'findQuery',
-    'getById',
-    'hasRecordForId',
-    'metaForType',
-    'modelFor',
-    'modelFactoryFor',
-    'normalize',
-    'push',
-    'pushMany',
-    'pushPayload',
-    'recordForId',
-    'recordIsLoaded',
-    'update',
-    'serializerFor',
-    'setMetadataFor',
-    'unloadAll'
-  ];
 
   recast.visit(ast, {
     visitCallExpression: function(path) {
       var node = path.node;
 
-      if (isRelationshipMacro(node)) {
+      if (emberData.isRelationshipMacro(node)) {
         replaceArguments(path);
       }
 
-      if (isStoreMethod(node)) {
+      if (emberData.isStoreMethod(node)) {
         replaceArguments(path);
       }
 
       this.traverse(path);
     }
   });
-
-  function isHasMany(node) {
-    var mightBeHasMany;
-
-    if (types.MemberExpression.check(node.callee)) {
-      mightBeHasMany = node.callee.object.name === 'DS' &&
-        node.callee.property.name === 'hasMany';
-    } else if (types.Identifier.check(node.callee)) {
-      mightBeHasMany = node.callee.name === 'hasMany';
-    }
-
-    return mightBeHasMany && isSupportedModelCall(node);
-  }
-
-  function isBelongsTo(node) {
-    var mightBeBelongsTo;
-
-    if (types.MemberExpression.check(node.callee)) {
-      mightBeBelongsTo = node.callee.object.name === 'DS' &&
-        node.callee.property.name === 'belongsTo';
-    } else if (types.Identifier.check(node.callee)) {
-      mightBeBelongsTo = node.callee.name === 'belongsTo';
-    }
-
-    return mightBeBelongsTo && isSupportedModelCall(node);
-  }
-
-  // First arguments is a string, or a variable that is capitalized, or a member expression.
-  function isSupportedModelCall(node) {
-    var firstArg;
-    if (node.arguments.length) {
-      firstArg = node.arguments[0];
-      if (types.Literal.check(firstArg)) {
-        return typeof firstArg.value === 'string';
-      } else if (types.Identifier.check(firstArg)) {
-        return firstArg.name[0] === firstArg.name[0].toUpperCase();
-      }
-      return types.MemberExpression.check(firstArg);
-    }
-    return false;
-  }
-
-  function isRelationshipMacro(node) {
-    return isHasMany(node) || isBelongsTo(node);
-  }
 
   function rebuildArguments(args) {
     var reargs        = args.slice();
@@ -132,16 +62,6 @@ module.exports = function transformEmberDataModelLookups(source){
     }
 
     return modelName;
-  }
-
-
-  function isStoreMethod(node) {
-    var maybeStoreMethod;
-    if (types.MemberExpression.check(node.callee)) {
-      maybeStoreMethod = node.callee.object.property && node.callee.object.property.name === 'store' &&
-        STORE_METHODS.indexOf(node.callee.property.name) >= 0;
-    }
-    return maybeStoreMethod && isSupportedModelCall(node);
   }
 
   return recast.print(ast, {tabWidth: 2, quote: 'single'}).code;
